Editable metadata schemas and initial support for organising assets in collections (#17)
This adds support for associating metadata with assets and organising them in collections.

This focuses mostly on the supporting infrastructure for displaying and editing schemas and validating assets. We add support for a single metadata type – free text.

Metadata is stored in a JSON field. It is not currently indexed – the intention is to add support for this once we implement searching and filtering. There's a fair bit to be decided about how we approach that – currently my thought is that we create sqlite expression indexes at runtime.