adds support for displaying html error messages in a modal (#64)
The API allows callers to launch a new window displaying some HTML (rendered from react) that presents information to the user. The caller then awaits a promise for the window to close, then returns.

In the process of doing this, we refactor the app entrypoint a bit to make support for different inital views a bit cleaner.

This potentially also interesting a first step towards a more multi-window UI, so interested in feedback on the API for 'modal' use cases (perhaps instead of providing some HTML to render, the caller could provide the path to a route to display).